RDMA/bnxt_re: Trim the QP mtu if it is higher than L2 MTU
Gen P5 chip allow sending roce packets bigger than L2 MTU.
But at the responder the packets are dropped if mismatch is
detected in the incoming packet size and the interface MTU.
If this is attempted by a malicious or badly written code
the chip drops incoming packets and corresponding rx error
counter is incremented. Since there is no response from
the responder QP the requester QP moves to ACK-Timeout
state.

Fixing the QP mtu setting to trim the qp-mtu to the lower
bound of L2 MTU during modify-qp from init to rtr.

Temporarily driver would fail the modify-request with mis-
matching MTU due to unavilability of a method which would
tell user qp about trimmed MTU.

Moved the mtu initialization code to separate inline
functions for ease of flow.

Signed-off-by: Bharat Gooty <bharat.gooty@broadcom.com>